requirements.txt
pandas
supabase
requests
python-dotenv
polars